refactored: icons dir
