+++ src/components/_icons/CellarGradientIcon.tsx
@@ -1,0 +1,25 @@
+import { Icon, IconProps } from '@chakra-ui/react'
+import { VFC } from 'react'
+
+export const CellarGradientIcon: VFC<IconProps> = props => (
+  <Icon viewBox='0 0 22 22' {...props}>
+    <path
+      d='M11 5.5C12.5209 5.5 13.75 4.2709 13.75 2.75C13.75 1.2291 12.5209 0 11 0C9.4791 0 8.25 1.2291 8.25 2.75C8.25 4.2709 9.4791 5.5 11 5.5ZM19.25 13.75C20.7709 13.75 22 12.5209 22 11C22 9.4791 20.7709 8.25 19.25 8.25C17.7291 8.25 16.5 9.4791 16.5 11C16.5 12.5209 17.7291 13.75 19.25 13.75ZM2.75 8.25C1.2291 8.25 0 9.48794 0 11C0 12.5209 1.2291 13.75 2.75 13.75C4.2709 13.75 5.5 12.5209 5.5 11C5.5 9.4791 4.2709 8.25 2.75 8.25ZM12.2556 11L19.135 4.12058L17.8883 2.87379L11.0088 9.75321L4.12942 2.87379L2.88264 4.12058L9.76206 11L2.88264 17.8794L4.12942 19.1262L11.0088 12.2468L17.8883 19.1262L19.135 17.8794L12.2556 11ZM11.0088 16.5C9.48794 16.5 8.25884 17.7291 8.25884 19.25C8.25884 20.7709 9.48794 22 11.0088 22C12.5297 22 13.7588 20.7709 13.7588 19.25C13.7588 17.7291 12.5297 16.5 11.0088 16.5Z'
+      fill='url(#paint0_linear_2736_39986)'
+    />
+    <defs>
+      <linearGradient
+        id='paint0_linear_2736_39986'
+        x1='0'
+        y1='11'
+        x2='22'
+        y2='11'
+        gradientUnits='userSpaceOnUse'
+      >
+        <stop stopColor='#E8D37C' />
+        <stop offset='0.5833' stopColor='#F6851B' />
+        <stop offset='1' stopColor='#EC3275' />
+      </linearGradient>
+    </defs>
+  </Icon>
+)
