+++ src/components/_icons/LogoIcon.tsx
@@ -1,0 +1,43 @@
+import { Icon, IconProps } from '@chakra-ui/react'
+import { VFC } from 'react'
+
+export const LogoIcon: VFC<IconProps> = props => (
+  <Icon viewBox='0 0 28 28' {...props}>
+    <path
+      d='M13.9316 5.83301C15.4712 5.83301 16.9478 5.21875 18.0366 4.125C19.125 3.03125 19.7368 1.54688 19.7368 0H8.12695C8.12695 0.765625 8.27686 1.51562 8.56006 2.20996C8.84912 2.91992 9.27734 3.57227 9.82715 4.125C10.916 5.21875 12.3921 5.83301 13.9316 5.83301Z'
+      fill='currentColor'
+    />
+    <path
+      d='M0 19.834V8.16699C1.53955 8.16699 3.01611 8.78125 4.10449 9.87598C5.19336 10.9697 5.80469 12.4531 5.80469 14C5.80469 15.5479 5.19336 17.0312 4.10449 18.125C3.01611 19.2188 1.53955 19.834 0 19.834Z'
+      fill='currentColor'
+    />
+    <path
+      d='M13.9316 8.16699C10.7261 8.16699 8.12695 10.7783 8.12695 14C8.12695 17.2217 10.7261 19.834 13.9316 19.834C17.1377 19.834 19.7368 17.2217 19.7368 14C19.7368 10.7783 17.1377 8.16699 13.9316 8.16699Z'
+      fill='currentColor'
+    />
+    <path
+      d='M22.0586 14C22.0586 15.5479 22.6704 17.0312 23.7588 18.125C24.8477 19.2188 26.3237 19.834 27.8633 19.834V8.16699C26.3237 8.16699 24.8477 8.78125 23.7588 9.87598C22.6704 10.9697 22.0586 12.4531 22.0586 14Z'
+      fill='currentColor'
+    />
+    <path
+      d='M27.8633 0V5.83301C26.3237 5.83301 24.8477 5.21875 23.7588 4.125C22.6704 3.03125 22.0586 1.54688 22.0586 0H27.8633Z'
+      fill='currentColor'
+    />
+    <path
+      d='M0 5.83301C1.53955 5.83301 3.01611 5.21875 4.10449 4.125C4.41943 3.80859 4.69434 3.45996 4.92578 3.08691C5.01807 2.93848 5.10303 2.78613 5.18115 2.62988C5.25244 2.4873 5.31836 2.34277 5.37793 2.19531C5.65723 1.50488 5.80469 0.759766 5.80469 0H0V5.83301Z'
+      fill='currentColor'
+    />
+    <path
+      d='M9.82715 23.876C10.916 22.7812 12.3921 22.167 13.9316 22.167C15.4712 22.167 16.9478 22.7812 18.0366 23.876C19.125 24.9697 19.7368 26.4531 19.7368 28H8.12695C8.12695 26.4531 8.73877 24.9697 9.82715 23.876Z'
+      fill='currentColor'
+    />
+    <path
+      d='M22.0586 28H27.8633V22.167C26.3237 22.167 24.8477 22.7812 23.7588 23.876C22.6704 24.9697 22.0586 26.4531 22.0586 28Z'
+      fill='currentColor'
+    />
+    <path
+      d='M0 28V22.167C1.53955 22.167 3.01611 22.7812 4.10449 23.876C5.19336 24.9697 5.80469 26.4531 5.80469 28H0Z'
+      fill='currentColor'
+    />
+  </Icon>
+)
