+++ src/components/_icons/LogoTextIcon.tsx
@@ -1,111 +1,6 @@
 import { Icon, IconProps } from '@chakra-ui/react'
 import { VFC } from 'react'
 
-export const OverviewIcon: VFC<IconProps> = props => (
-  <Icon viewBox='0 0 20 20' {...props}>
-    <path
-      d='M6.19718 3.09859C5.34507 2.24648 3.95775 2.24648 3.09859 3.09859C2.24648 3.95775 2.24648 5.33803 3.09859 6.19718C3.95775 7.0493 5.33803 7.0493 6.19718 6.19718C7.0493 5.33803 7.0493 3.95775 6.19718 3.09859ZM13.8028 16.9014C14.662 17.7535 16.0423 17.7535 16.9014 16.9014C17.7535 16.0423 17.7535 14.662 16.9014 13.8028C16.0493 12.9437 14.662 12.9437 13.8028 13.8028C12.9437 14.662 12.9507 16.0423 13.8028 16.9014ZM0 10.7042V20H9.29577V10.7042H0ZM7.88732 12.1127V18.5915H1.40845V12.1127H7.88732ZM20 10.7042H10.7042V20H20V10.7042ZM12.1127 18.5915V12.1127H18.5915V18.5915H12.1127ZM9.29577 0H0V9.29577H9.29577V0ZM1.40845 1.40845H7.88732V7.88732H1.40845V1.40845ZM20 0H10.7042V9.29577H20V0ZM12.1127 1.40845H18.5915V7.88732H12.1127V1.40845Z'
-      fill='currentColor'
-    />
-  </Icon>
-)
-
-export const OverviewGradientIcon: VFC<IconProps> = props => (
-  <Icon viewBox='0 0 20 20' {...props}>
-    <path
-      d='M6.19718 3.09859C5.34507 2.24648 3.95775 2.24648 3.09859 3.09859C2.24648 3.95775 2.24648 5.33803 3.09859 6.19718C3.95775 7.0493 5.33803 7.0493 6.19718 6.19718C7.0493 5.33803 7.0493 3.95775 6.19718 3.09859ZM13.8028 16.9014C14.662 17.7535 16.0423 17.7535 16.9014 16.9014C17.7535 16.0423 17.7535 14.662 16.9014 13.8028C16.0493 12.9437 14.662 12.9437 13.8028 13.8028C12.9437 14.662 12.9507 16.0423 13.8028 16.9014ZM0 10.7042V20H9.29577V10.7042H0ZM7.88732 12.1127V18.5915H1.40845V12.1127H7.88732ZM20 10.7042H10.7042V20H20V10.7042ZM12.1127 18.5915V12.1127H18.5915V18.5915H12.1127ZM9.29577 0H0V9.29577H9.29577V0ZM1.40845 1.40845H7.88732V7.88732H1.40845V1.40845ZM20 0H10.7042V9.29577H20V0ZM12.1127 1.40845H18.5915V7.88732H12.1127V1.40845Z'
-      fill='url(#paint0_linear_2736_9926)'
-    />
-    <defs>
-      <linearGradient
-        id='paint0_linear_2736_9926'
-        x1='0'
-        y1='10'
-        x2='20'
-        y2='10'
-        gradientUnits='userSpaceOnUse'
-      >
-        <stop stopColor='#E8D37C' />
-        <stop offset='0.5833' stopColor='#F6851B' />
-        <stop offset='1' stopColor='#EC3275' />
-      </linearGradient>
-    </defs>
-  </Icon>
-)
-
-export const CellarIcon: VFC<IconProps> = props => (
-  <Icon viewBox='0 0 22 22' {...props}>
-    <path
-      d='M11 5.5C12.5209 5.5 13.75 4.2709 13.75 2.75C13.75 1.2291 12.5209 0 11 0C9.4791 0 8.25 1.2291 8.25 2.75C8.25 4.2709 9.4791 5.5 11 5.5ZM19.25 13.75C20.7709 13.75 22 12.5209 22 11C22 9.4791 20.7709 8.25 19.25 8.25C17.7291 8.25 16.5 9.4791 16.5 11C16.5 12.5209 17.7291 13.75 19.25 13.75ZM2.75 8.25C1.2291 8.25 0 9.48794 0 11C0 12.5209 1.2291 13.75 2.75 13.75C4.2709 13.75 5.5 12.5209 5.5 11C5.5 9.4791 4.2709 8.25 2.75 8.25ZM12.2556 11L19.135 4.12058L17.8883 2.87379L11.0088 9.75321L4.12942 2.87379L2.88264 4.12058L9.76206 11L2.88264 17.8794L4.12942 19.1262L11.0088 12.2468L17.8883 19.1262L19.135 17.8794L12.2556 11ZM11.0088 16.5C9.48794 16.5 8.25884 17.7291 8.25884 19.25C8.25884 20.7709 9.48794 22 11.0088 22C12.5297 22 13.7588 20.7709 13.7588 19.25C13.7588 17.7291 12.5297 16.5 11.0088 16.5Z'
-      fill='currentColor'
-    />
-  </Icon>
-)
-
-export const CellarGradientIcon: VFC<IconProps> = props => (
-  <Icon viewBox='0 0 22 22' {...props}>
-    <path
-      d='M11 5.5C12.5209 5.5 13.75 4.2709 13.75 2.75C13.75 1.2291 12.5209 0 11 0C9.4791 0 8.25 1.2291 8.25 2.75C8.25 4.2709 9.4791 5.5 11 5.5ZM19.25 13.75C20.7709 13.75 22 12.5209 22 11C22 9.4791 20.7709 8.25 19.25 8.25C17.7291 8.25 16.5 9.4791 16.5 11C16.5 12.5209 17.7291 13.75 19.25 13.75ZM2.75 8.25C1.2291 8.25 0 9.48794 0 11C0 12.5209 1.2291 13.75 2.75 13.75C4.2709 13.75 5.5 12.5209 5.5 11C5.5 9.4791 4.2709 8.25 2.75 8.25ZM12.2556 11L19.135 4.12058L17.8883 2.87379L11.0088 9.75321L4.12942 2.87379L2.88264 4.12058L9.76206 11L2.88264 17.8794L4.12942 19.1262L11.0088 12.2468L17.8883 19.1262L19.135 17.8794L12.2556 11ZM11.0088 16.5C9.48794 16.5 8.25884 17.7291 8.25884 19.25C8.25884 20.7709 9.48794 22 11.0088 22C12.5297 22 13.7588 20.7709 13.7588 19.25C13.7588 17.7291 12.5297 16.5 11.0088 16.5Z'
-      fill='url(#paint0_linear_2736_39986)'
-    />
-    <defs>
-      <linearGradient
-        id='paint0_linear_2736_39986'
-        x1='0'
-        y1='11'
-        x2='22'
-        y2='11'
-        gradientUnits='userSpaceOnUse'
-      >
-        <stop stopColor='#E8D37C' />
-        <stop offset='0.5833' stopColor='#F6851B' />
-        <stop offset='1' stopColor='#EC3275' />
-      </linearGradient>
-    </defs>
-  </Icon>
-)
-
-export const LogoIcon: VFC<IconProps> = props => (
-  <Icon viewBox='0 0 28 28' {...props}>
-    <path
-      d='M13.9316 5.83301C15.4712 5.83301 16.9478 5.21875 18.0366 4.125C19.125 3.03125 19.7368 1.54688 19.7368 0H8.12695C8.12695 0.765625 8.27686 1.51562 8.56006 2.20996C8.84912 2.91992 9.27734 3.57227 9.82715 4.125C10.916 5.21875 12.3921 5.83301 13.9316 5.83301Z'
-      fill='currentColor'
-    />
-    <path
-      d='M0 19.834V8.16699C1.53955 8.16699 3.01611 8.78125 4.10449 9.87598C5.19336 10.9697 5.80469 12.4531 5.80469 14C5.80469 15.5479 5.19336 17.0312 4.10449 18.125C3.01611 19.2188 1.53955 19.834 0 19.834Z'
-      fill='currentColor'
-    />
-    <path
-      d='M13.9316 8.16699C10.7261 8.16699 8.12695 10.7783 8.12695 14C8.12695 17.2217 10.7261 19.834 13.9316 19.834C17.1377 19.834 19.7368 17.2217 19.7368 14C19.7368 10.7783 17.1377 8.16699 13.9316 8.16699Z'
-      fill='currentColor'
-    />
-    <path
-      d='M22.0586 14C22.0586 15.5479 22.6704 17.0312 23.7588 18.125C24.8477 19.2188 26.3237 19.834 27.8633 19.834V8.16699C26.3237 8.16699 24.8477 8.78125 23.7588 9.87598C22.6704 10.9697 22.0586 12.4531 22.0586 14Z'
-      fill='currentColor'
-    />
-    <path
-      d='M27.8633 0V5.83301C26.3237 5.83301 24.8477 5.21875 23.7588 4.125C22.6704 3.03125 22.0586 1.54688 22.0586 0H27.8633Z'
-      fill='currentColor'
-    />
-    <path
-      d='M0 5.83301C1.53955 5.83301 3.01611 5.21875 4.10449 4.125C4.41943 3.80859 4.69434 3.45996 4.92578 3.08691C5.01807 2.93848 5.10303 2.78613 5.18115 2.62988C5.25244 2.4873 5.31836 2.34277 5.37793 2.19531C5.65723 1.50488 5.80469 0.759766 5.80469 0H0V5.83301Z'
-      fill='currentColor'
-    />
-    <path
-      d='M9.82715 23.876C10.916 22.7812 12.3921 22.167 13.9316 22.167C15.4712 22.167 16.9478 22.7812 18.0366 23.876C19.125 24.9697 19.7368 26.4531 19.7368 28H8.12695C8.12695 26.4531 8.73877 24.9697 9.82715 23.876Z'
-      fill='currentColor'
-    />
-    <path
-      d='M22.0586 28H27.8633V22.167C26.3237 22.167 24.8477 22.7812 23.7588 23.876C22.6704 24.9697 22.0586 26.4531 22.0586 28Z'
-      fill='currentColor'
-    />
-    <path
-      d='M0 28V22.167C1.53955 22.167 3.01611 22.7812 4.10449 23.876C5.19336 24.9697 5.80469 26.4531 5.80469 28H0Z'
-      fill='currentColor'
-    />
-  </Icon>
-)
-
 export const LogoTextIcon: VFC<IconProps> = props => (
   <Icon viewBox='0 0 162 28' {...props}>
     <path
