+++ src/components/_icons/OverviewGradientIcon.tsx
@@ -1,0 +1,25 @@
+import { Icon, IconProps } from '@chakra-ui/react'
+import { VFC } from 'react'
+
+export const OverviewGradientIcon: VFC<IconProps> = props => (
+  <Icon viewBox='0 0 20 20' {...props}>
+    <path
+      d='M6.19718 3.09859C5.34507 2.24648 3.95775 2.24648 3.09859 3.09859C2.24648 3.95775 2.24648 5.33803 3.09859 6.19718C3.95775 7.0493 5.33803 7.0493 6.19718 6.19718C7.0493 5.33803 7.0493 3.95775 6.19718 3.09859ZM13.8028 16.9014C14.662 17.7535 16.0423 17.7535 16.9014 16.9014C17.7535 16.0423 17.7535 14.662 16.9014 13.8028C16.0493 12.9437 14.662 12.9437 13.8028 13.8028C12.9437 14.662 12.9507 16.0423 13.8028 16.9014ZM0 10.7042V20H9.29577V10.7042H0ZM7.88732 12.1127V18.5915H1.40845V12.1127H7.88732ZM20 10.7042H10.7042V20H20V10.7042ZM12.1127 18.5915V12.1127H18.5915V18.5915H12.1127ZM9.29577 0H0V9.29577H9.29577V0ZM1.40845 1.40845H7.88732V7.88732H1.40845V1.40845ZM20 0H10.7042V9.29577H20V0ZM12.1127 1.40845H18.5915V7.88732H12.1127V1.40845Z'
+      fill='url(#paint0_linear_2736_9926)'
+    />
+    <defs>
+      <linearGradient
+        id='paint0_linear_2736_9926'
+        x1='0'
+        y1='10'
+        x2='20'
+        y2='10'
+        gradientUnits='userSpaceOnUse'
+      >
+        <stop stopColor='#E8D37C' />
+        <stop offset='0.5833' stopColor='#F6851B' />
+        <stop offset='1' stopColor='#EC3275' />
+      </linearGradient>
+    </defs>
+  </Icon>
+)
