+++ src/components/_icons/index.tsx
@@ -1,0 +1,6 @@
+export { LogoIcon } from './LogoIcon'
+export { LogoTextIcon } from './LogoTextIcon'
+export { CellarIcon } from './CellarIcon'
+export { CellarGradientIcon } from './CellarGradientIcon'
+export { OverviewIcon } from './OverviewIcon'
+export { OverviewGradientIcon } from './OverviewGradientIcon'
